Include MAC by default if encryption supports it
DEVSIX-8622

+++ kernel/src/main/java/com/itextpdf/kernel/crypto/securityhandler/StandardHandlerUsingStandard40.java
@@ -171,7 +171,7 @@ public class StandardHandlerUsingStandard40 extends StandardSecurityHandler {
         byte[] oValue = getIsoBytes(encryptionDictionary.getAsString(PdfName.O));
 
         PdfNumber pValue = (PdfNumber) encryptionDictionary.get(PdfName.P);
-        this.permissions = pValue.longValue();
+        this.permissions = pValue.intValue();
 
         this.documentId = documentId;
         keyLength = getKeyLength(encryptionDictionary);
